backend1/requirements.txt
flask
flask_cors
requests
boto3
PyJWT[crypto]
aws-lambda-wsgi